Refactor UnifiedAI into package
