+++ UnifiedAI.md
@@ -1,7 +1,5 @@
 # UnifiedAI
 
-=======
-=======
 `UnifiedAI` is a minimal async engine that relies on Redis for short term storage.
 It exposes a `connect()` method to establish the connection and a `close()`
 coroutine to gracefully shut it down.
@@ -22,9 +20,7 @@ The provided context manager automatically invokes `engine.close()` after the
 `yield`, ensuring Redis resources are released. When not using the context
 manager, make sure to `await engine.close()` yourself during application
 shutdown to avoid open connections.
-=======
 `yield`, ensuring Redis resources are released.
-=======
 UnifiedAI is a modular architecture that combines four engines to provide empathetic interactions, reasoning, high‑speed data processing and ethical oversight.  It exposes a small FastAPI service for demonstration purposes.
 
 ## Components and Intent
@@ -33,8 +29,6 @@ UnifiedAI is a modular architecture that combines four engines to provide empath
 - **BrainEngine** – stores memories in SQLite, applies rule-based reasoning, and learns from interactions.
 - **OpticalEngine** – publishes events through Redis with optional networking features like smart packet shaping.
 - **AuraEngine** – validates output against rules loaded from `ethics_rules.json` for ethical compliance.
-=======
-=======
 - **SoulEngine** – detects sentiment in user messages and crafts human‑like replies.
 - **BrainEngine** – stores memories in SQLite and performs simple reasoning based on past interactions.
 - **OpticalEngine** – communicates through Redis channels to broadcast events.
@@ -51,14 +45,12 @@ The orchestrator coordinates these engines so that input flows through AuraEngin
 2. Start a local Redis server (required for the OpticalEngine).
 3. Run the API server:
    ```bash
-   python unifiedai.py
+   python -m unified_ai
    ```
 4. Interact with the system using `curl` or any HTTP client:
    ```bash
    curl -X POST "http://localhost:8000/query" \
         -H "Content-Type: application/json" -d '{"query": "Hello there"}'
-=======
-=======
    curl -X POST -H "Content-Type: application/json" \
         -d '{"message": "Hello there"}' http://localhost:8000/chat
    ```
@@ -281,5 +273,3 @@ future integrations. Example entries include:
 Features are disabled by default. Enable them using the API or by calling
 `UnifiedAI.enable_feature()` with the feature key from `NETWORK_FEATURES`.
 
-=======
-=======
