+++ requirements.txt
@@ -1,23 +1,8 @@
-transformers
-torch
-Pillow
-requests
-openai
-textblob
-redis
-pytest
-pytest-asyncio
-=======
-=======
-=======
-redis
-pytest
-pytest-asyncio
-=======
 fastapi
 aiosqlite
 redis
 uvicorn
 pydantic
-=======
-=======
+textblob
+pytest
+pytest-asyncio
