+++ unified_ai/network_features.py
@@ -1,281 +1,4 @@
-import asyncio
-import json
 import logging
-import signal
-from typing import Any, Optional, AsyncGenerator
-=======
-=======
-import logging
-from typing import Optional, AsyncGenerator
-
-from fastapi import FastAPI, HTTPException
-from pydantic import BaseModel
-from contextlib import asynccontextmanager
-import redis.asyncio as redis
-import aiosqlite
-from transformers import pipeline
-=======
-=======
-from textblob import TextBlob
-
-
-class SoulEngine:
-    """Handle empathetic interactions and emotional analysis."""
-
-    def __init__(self) -> None:
-        self.logger = logging.getLogger(self.__class__.__name__)
-        # pragma: no cover - model download can be slow
-        self.classifier = pipeline("text-classification", model="distilbert-base-uncased-emotion")
-
-    async def analyze_emotion(self, text: str) -> str:
-        """Return a nuanced emotion label for the text."""
-        try:
-            result = self.classifier(text)[0]
-            return result["label"].lower()
-        except Exception as exc:  # pragma: no cover - defensive
-            self.logger.error("Emotion analysis failed: %s", exc)
-            return "neutral"
-
-    async def craft_reply(self, text: str, emotion: str) -> str:
-        """Craft an empathetic reply based on detected emotion."""
-        responses = {
-            "joy": "That's wonderful!",
-            "sadness": "I'm here for you.",
-            "anger": "I understand your frustration.",
-            "fear": "That sounds scary.",
-            "surprise": "That's surprising!",
-            "neutral": "I see.",
-        }
-        base = responses.get(emotion, "I see.")
-=======
-=======
-
-    async def analyze_emotion(self, text: str) -> str:
-        """Return a basic emotion label for the text."""
-        try:
-            polarity = TextBlob(text).sentiment.polarity
-            if polarity > 0.2:
-                return "positive"
-            if polarity < -0.2:
-                return "negative"
-            return "neutral"
-        except Exception as exc:  # pragma: no cover - defensive
-            self.logger.error("Emotion analysis failed: %s", exc)
-            return "unknown"
-
-    async def craft_reply(self, text: str, emotion: str) -> str:
-        """Craft a simple empathetic reply."""
-        base = {
-            "positive": "I'm glad to hear that!",
-            "negative": "I'm sorry to hear that.",
-            "neutral": "I understand.",
-        }.get(emotion, "I see.")
- main
-        return f"{base} You said: {text}"
-
-
-class BrainEngine:
-    """Reasoning, memory management, and learning using SQLite."""
-
-    def __init__(self, db_path: str = "brain.db") -> None:
-        self.db_path = db_path
-        self.logger = logging.getLogger(self.__class__.__name__)
-
-    async def initialize(self) -> None:
-        self.db = await aiosqlite.connect(self.db_path)
-        await self.db.execute(
-            "CREATE TABLE IF NOT EXISTS memories (key TEXT PRIMARY KEY, content TEXT, timestamp TEXT, access_count INTEGER)"
-        )
-        await self.db.commit()
-        self.rules = {
-            "productivity": "Try time-blocking and prioritizing tasks with a Pomodoro technique.",
-            "learning": "Consider spaced repetition and hands-on projects.",
-        }
-
-    async def store_memory(self, key: str, value: str) -> None:
-        try:
-            await self.db.execute(
-                "INSERT OR REPLACE INTO memories (key, content, timestamp, access_count) VALUES (?, ?, datetime('now'), COALESCE((SELECT access_count FROM memories WHERE key = ?),0))",
-                (key, value, key),
-            )
-            await self.db.commit()
-=======
-=======
-        async with aiosqlite.connect(self.db_path) as db:
-            await db.execute(
-                "CREATE TABLE IF NOT EXISTS memories (key TEXT PRIMARY KEY, value TEXT)"
-            )
-            await db.commit()
-
-    async def store_memory(self, key: str, value: str) -> None:
-        try:
-            async with aiosqlite.connect(self.db_path) as db:
-                await db.execute(
-                    "REPLACE INTO memories (key, value) VALUES (?, ?)",
-                    (key, value),
-                )
-                await db.commit()
-        except Exception as exc:
-            self.logger.error("Storing memory failed: %s", exc)
-
-    async def retrieve_memory(self, key: str) -> Optional[str]:
-        try:
-            async with self.db.execute("SELECT content, access_count FROM memories WHERE key = ?", (key,)) as cursor:
-                row = await cursor.fetchone()
-                if row:
-                    await self.db.execute("UPDATE memories SET access_count = access_count + 1 WHERE key = ?", (key,))
-                    await self.db.commit()
-                    return row[0]
-=======
-=======
-            async with aiosqlite.connect(self.db_path) as db:
-                async with db.execute(
-                    "SELECT value FROM memories WHERE key = ?", (key,)
-                ) as cursor:
-                    row = await cursor.fetchone()
-                    if row:
-                        return row[0]
-            return None
-        except Exception as exc:
-            self.logger.error("Retrieving memory failed: %s", exc)
-            return None
-
-    async def reason(self, text: str) -> str:
-        lowered = text.lower()
-        for key, val in self.rules.items():
-            if key in lowered:
-                await self.learn(text)
-                return f"Reasoned: {val}"
-        mem = await self.retrieve_memory(lowered)
-        if mem:
-            return f"I recall you said: {mem}"
-        await self.store_memory(lowered, text)
-        return f"Reasoned: {lowered}"
-
-    async def learn(self, item: str) -> bool:
-        try:
-            key = f"item_{len(await self.memory_count())}"
-            await self.store_memory(key, item)
-            return True
-        except Exception as exc:
-            self.logger.error("Learning failed: %s", exc)
-            return False
-
-    async def memory_count(self) -> int:
-        async with self.db.execute("SELECT COUNT(*) FROM memories") as cursor:
-            row = await cursor.fetchone()
-            return row[0] if row else 0
-
-    async def close(self) -> None:
-        await self.db.close()
-=======
-=======
-        """Very simple reasoning: echo known memories or store new."""
-        mem = await self.retrieve_memory(text)
-        if mem:
-            return f"I recall you said: {mem}"
-        await self.store_memory(text, text)
-        return "Thanks for telling me."        
-
-
-class OpticalEngine:
-    """High-speed data processing and communication via Redis."""
-
-    def __init__(self, redis_client: redis.Redis, feature_manager: "NetworkFeatureManager") -> None:
-        self.redis = redis_client
-        self.features = feature_manager
-        self.logger = logging.getLogger(self.__class__.__name__)
-
-    async def initialize(self) -> None:
-        """Enable important networking features."""
-        for feat in [
-            "optical_path_diversity_planner",
-            "real_time_bfd_path_scoring",
-            "smart_packet_shaping",
-        ]:
-            self.features.enable(feat)
-        self.logger.info("OpticalEngine enabled: %s", self.features.list_enabled())
-
-    async def transfer_data(self, data: Any, target: str) -> bool:
-        """Publish data to a Redis channel with optional feature logic."""
-        try:
-            if "smart_packet_shaping" in self.features.enabled:
-                self.logger.debug("Applying smart packet shaping")
-            await self.redis.publish(target, str(data))
-            return True
-        except Exception as exc:
-            self.logger.error("Publish failed: %s", exc)
-            return False
-=======
-=======
-    def __init__(self, url: str = "redis://localhost:6379/0") -> None:
-        self.redis = redis.from_url(url)
-        self.logger = logging.getLogger(self.__class__.__name__)
-
-    async def publish(self, channel: str, message: str) -> None:
-        try:
-            await self.redis.publish(channel, message)
-        except Exception as exc:
-            self.logger.error("Publish failed: %s", exc)
-
-    async def subscribe(self, channel: str) -> AsyncGenerator[str, None]:
-        pubsub = self.redis.pubsub()
-        await pubsub.subscribe(channel)
-        try:
-            async for item in pubsub.listen():
-                if item.get("type") == "message":
-                    yield item.get("data")
-        finally:
-            await pubsub.unsubscribe(channel)
-
-    async def process_message(self, message: str) -> None:
-        """Placeholder for future message handling."""
-        self.logger.info("Received message: %s", message)
-
-=======
-=======
-
-class AuraEngine:
-    """Ethical oversight and contextual awareness."""
-
-    def __init__(self, rules_file: str = "ethics_rules.json") -> None:
-        self.rules_file = rules_file
-        self.logger = logging.getLogger(self.__class__.__name__)
-        self.ethics_rules = {
-            "no_harm": "Do not generate harmful content",
-            "no_bias": "Avoid biased or discriminatory language",
-            "privacy": "Protect user data and privacy",
-        }
-        self.load_rules()
-
-    def load_rules(self) -> None:
-        try:
-            with open(self.rules_file, "r", encoding="utf-8") as f:
-                self.ethics_rules.update(json.load(f))
-        except FileNotFoundError:
-            self.logger.warning("Ethics rules file not found, using defaults")
-        except Exception as exc:  # pragma: no cover - defensive
-            self.logger.error("Failed to load ethics rules: %s", exc)
-
-    async def validate_output(self, output: str) -> bool:
-        lowered = output.lower()
-        if "harm" in lowered and "no_harm" in self.ethics_rules:
-            return False
-        if any(word in lowered for word in ["bias", "discriminate"]):
-=======
-=======
-    def __init__(self) -> None:
-        self.forbidden = {"hate", "kill", "malicious"}
-        self.logger = logging.getLogger(self.__class__.__name__)
-
-    async def check(self, text: str) -> bool:
-        """Return True if text passes ethical check."""
-        lowered = text.lower()
-        if any(word in lowered for word in self.forbidden):
-            self.logger.warning("Blocked unethical input: %s", text)
- main
-            return False
-        return True
 
 
 NETWORK_FEATURES = {
@@ -458,6 +181,7 @@ NETWORK_FEATURES = {
     "geo_fencing_firewall_zones": "Geo-fencing firewall zones \u2013 physical limits",
     "immutable_telemetry_chain": "Immutable telemetry chain \u2013 forensics",
     "taint_tracking_in_flows": "Taint-tracking in flows \u2013 data lineage",
+    "taint_tracking_in_flows": "Taint-tracking in flows \u2013 data lineage",
     "topology_chaos_sim_tool": "Topology chaos sim tool \u2013 disaster drill",
     "kernel_pinned_path_cache": "Kernel-pinned path cache \u2013 route lock",
     "protocol_aware_zeroing_buffer": "Protocol-aware zeroing buffer \u2013 protect remnants",
@@ -482,6 +206,7 @@ NETWORK_FEATURES = {
 }
 
 
+
 class NetworkFeatureManager:
     """Manage optional networking features for UnifiedAI."""
 
@@ -502,156 +227,3 @@ class NetworkFeatureManager:
 
     def list_enabled(self) -> list[str]:
         return [NETWORK_FEATURES[f] for f in self.enabled]
-
-
-=======
-=======
-class UnifiedAI:
-    """Central orchestrator coordinating all engines."""
-
-    def __init__(self) -> None:
-        self.feature_manager = NetworkFeatureManager()
-        self.redis = redis.Redis(host="redis", port=6379, decode_responses=True)
-        self.soul = SoulEngine()
-        self.brain = BrainEngine()
-        self.optical = OpticalEngine(self.redis, self.feature_manager)
-        self.aura = AuraEngine()
-        self.logger = logging.getLogger(self.__class__.__name__)
-        self.bg_tasks: list[asyncio.Task] = []
-
-    async def initialize(self) -> None:
-        await self.brain.initialize()
-        await self.optical.initialize()
-        self.bg_tasks.append(asyncio.create_task(self._listener()))
-
-    async def close(self) -> None:
-        for t in self.bg_tasks:
-            t.cancel()
-        await self.redis.close()
-        await self.brain.close()
-
-    async def enable_feature(self, feature: str) -> None:
-        """Enable an optional networking feature."""
-        self.feature_manager.enable(feature)
-
-    def list_enabled_features(self) -> list[str]:
-        """Return descriptions of currently enabled networking features."""
-        return self.feature_manager.list_enabled()
-
-    async def interact(self, text: str) -> str:
-        if not await self.aura.validate_output(text):
-            raise HTTPException(status_code=400, detail="Inappropriate content")
-        emotion = await self.soul.analyze_emotion(text)
-        memory_response = await self.brain.reason(text)
-        await self.optical.transfer_data(memory_response, "UnifiedAI")
-        if await self.aura.validate_output(memory_response):
-            return await self.soul.craft_reply(memory_response, emotion)
-        return "Output blocked due to ethics rules"
-
-    async def _listener(self) -> None:
-        async for message in self.optical.subscribe("UnifiedAI"):
-            await self.optical.process_message(message)
-
-
-logging.basicConfig(level=logging.INFO)
-
-@asynccontextmanager
-async def lifespan(app: FastAPI):
-    ai = UnifiedAI()
-    await ai.initialize()
-    app.state.unified_ai = ai
-    ai.logger.info("Available networking features: %d", len(NETWORK_FEATURES))
-    try:
-        yield
-    finally:
-        await ai.close()
-=======
-=======
-        self.soul = SoulEngine()
-        self.brain = BrainEngine()
-        self.optical = OpticalEngine()
-        self.aura = AuraEngine()
-        self.logger = logging.getLogger(self.__class__.__name__)
-
-    async def setup(self) -> None:
-        await self.brain.initialize()
-
-    async def interact(self, text: str) -> str:
-        if not await self.aura.check(text):
-            raise HTTPException(status_code=400, detail="Inappropriate content")
-        emotion = await self.soul.analyze_emotion(text)
-        memory_response = await self.brain.reason(text)
-        await self.optical.publish("unifiedai_log", text)
-        reply = await self.soul.craft_reply(memory_response, emotion)
-        return reply
-
-
-logging.basicConfig(level=logging.INFO)
-engine = UnifiedAI()
-
-@asynccontextmanager
-async def lifespan(app: FastAPI):
-    """Handle startup and shutdown events for the app."""
-    await engine.setup()
-    yield
-    # No specific shutdown logic but placeholder for future cleanup
-
-app = FastAPI(lifespan=lifespan)
-
-
-class QueryRequest(BaseModel):
-    query: str
-
-
-
-
-@app.post("/query")
-async def query_endpoint(data: QueryRequest):
-    ai: UnifiedAI = app.state.unified_ai
-    try:
-        response = await ai.interact(data.query)
-        return {"response": response}
-    except HTTPException as exc:
-        raise exc
-    except Exception as exc:
-        raise HTTPException(status_code=500, detail=str(exc))
-
-@app.get("/health")
-async def health():
-    ai: UnifiedAI = app.state.unified_ai
-    try:
-        redis_status = "connected" if await ai.redis.ping() else "down"
-    except Exception:
-        redis_status = "down"
-    return {
-        "status": "healthy",
-        "redis": redis_status,
-        "features": ai.list_enabled_features(),
-    }
-
-@app.get("/metrics")
-async def metrics():
-    ai: UnifiedAI = app.state.unified_ai
-    count = await ai.brain.memory_count()
-    return {
-        "memory_count": count,
-        "enabled_network_features": ai.list_enabled_features(),
-    }
-=======
-=======
-class Interaction(BaseModel):
-    message: str
-
-
-
-
-@app.post("/chat")
-async def chat(data: Interaction) -> dict:
-    response = await engine.interact(data.message)
-    return {"response": response}
-
-
-if __name__ == "__main__":
-    import uvicorn
-
-    uvicorn.run(app, host="0.0.0.0", port=8000)
